requirements.txt
alabaster>=0.7.12
docutils>=0.15.2
ipython>=7.13.0
jinja2~=2.11.1
# matplotlib~=3.2.1
nose~=1.3.7
numpydoc~=0.9.2
prompt_toolkit>=3.0.5
pygments>=2.6.1
pynvim>=0.4.1
pytest>=5.4.1
pyyaml~=5.3.1
requests>=2.23.0
setuptools>=46.1.3
sphinx>=3.0.0
